ref: bug solved

+++ ewm-stat-service/common-dto-storage/src/main/java/ru/practicum/ewm/dto/StatisticCount.java
@@ -5,5 +5,5 @@ public interface StatisticCount {
 
     String getUri();
 
-    Integer getHits();
+    Long getHits();
 }
